fix: :fire: inject a correct type
