+++ package.json
@@ -6,7 +6,7 @@
   "main": "dist/sanitize.js",
   "types": "dist/sanitize.d.ts",
   "scripts": {
-    "build": "rimraf dist && npx tsc && cp src/type.d.ts dist",
+    "build": "rimraf dist && npx tsc",
     "release": "standard-version && git push --follow-tags origin main && yarn publish",
     "prepublish": "yarn build"
   },
@@ -52,6 +52,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.15.6",
     "@babel/plugin-transform-runtime": "^7.15.0",
     "@babel/preset-env": "^7.15.6",
+    "@nuxt/types": "^2.15.8",
     "@types/sanitize-html": "^2.6.2",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
@@ -60,11 +61,13 @@
     "babel-preset-minify": "^0.5.1",
     "eslint": "^8.5.0",
     "eslint-plugin-vue": "^8.2.0",
+    "nuxt": "^2.15.8",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
     "standard-version": "^9.3.1",
     "typescript": "^4.5.4",
-    "vue": "^2.6.14"
+    "vue": "^2.6.14",
+    "vuex": "^3.6.2"
   },
   "publishConfig": {
     "access": "public"
