+++ tsconfig.json
@@ -11,7 +11,8 @@
     "typeRoots": [
       "./node_modules/@types"
     ],
-    "sourceMap": true,
+    "types": ["@nuxt/types"],
+    "sourceMap": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "**/__tests__/*", "coverage"]
